Tried training the model with binary cross entropy... but still didn't work.
It still was focusing too heavy on non-car when the loss weights were 0.25 and 0.75
for the no_car and car class, respectively. This was still true even when trying with
a super imbalanced loss weights of 1 and 10 to really focus on car class. It still kept
hitting 75% accuracy (i.e. guessing pretty much everything as no_car since ~75% of the
dataset is no_car)